pyproject.toml
[tool.poetry]
name = "courseworkmodule7"
version = "0.1.0"
description = ""
authors = ["Andrey Grishanov <gau@1jz.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"
requests = "^2.31.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
djangorestframework-simplejwt = "^5.2.2"
coverage = "^7.2.7"
drf-yasg = "^1.21.7"
celery = "^5.3.1"
django-celery-beat = "^2.5.0"
redis = "^4.6.0"
django-cors-headers = "^4.2.0"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
